# REST/spec/swagger.yaml
swagger: '2.0'
info:
  version: '1.0'
  title: Gamifikator
schemes:
  - http
host: localhost
basePath: /v1
paths:
  /rules:
    get:
      tags:
        - rules
      summary: Get rules
      description: Returns list of rules
      operationId: getRules
      parameters:
        - in: query
          name: apitoken
          required: true
          type: string
      produces:
        - application/json
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Rule'
        '500':
          description: Internal server error
    post:
      tags:
        - rules
      summary: Add a new rule
      description: ''
      operationId: createRule
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: New rule
          required: true
          schema:
            $ref: '#/definitions/Rule'
      responses:
        '201':
            description: 'The rule has been created'
            schema:
              $ref: '#/definitions/Rule'
        '400':
          description: Invalid input
        '500':
          description: Internal server error
    put:
      tags:
        - rules
      summary: Update exisiting rule
      description: ''
      operationId: updateRule
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: New rule
          required: true
          schema:
            $ref: '#/definitions/Rule'
      responses:
        '201':
            description: 'The rule has been updated'
            schema:
              $ref: '#/definitions/Rule'
        '400':
          description: Invalid input
        '500':
          description: Internal server error
    delete:
      tags:
        - rules
      summary: Delete rule
      description: 'Delete rule'
      operationId: deleteRule
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Rule to be deleted
          required: true
          schema:
            $ref: '#/definitions/Rule'
      responses:
        '201':
            description: 'The rule has been deleted'
            schema:
              $ref: '#/definitions/Rule'
        '400':
          description: Invalid input
        '500':
          description: Internal server error
          
          
  /badges:
    get:
      tags:
        - badges
      summary: Get badges
      description: Returns list of badges
      operationId: getBadges
      produces:
        - application/json
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Badge'
        '500':
          description: Internal server error
    post:
      tags:
        - badges
      summary: Add a new badge
      description: ''
      operationId: createBadge
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: New badge
          required: true
          schema:
            $ref: '#/definitions/Badge'
      responses:
        '201':
            description: 'The badge has been created'
            schema:
              $ref: '#/definitions/Badge'
        '400':
          description: Invalid input
        '500':
          description: Internal server error
    put:
      tags:
        - badges
      summary: Update exisiting badge
      description: ''
      operationId: updateBadge
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Update badge
          required: true
          schema:
            $ref: '#/definitions/Badge'
      responses:
        '201':
            description: 'The badge has been updated'
            schema:
              $ref: '#/definitions/Badge'
        '400':
          description: Invalid input
        '500':
          description: Internal server error
    delete:
      tags:
        - badges
      summary: Delete badge
      description: 'Delete badge'
      operationId: deleteBadge
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Badge to be deleted
          required: true
          schema:
            $ref: '#/definitions/Badge'
      responses:
        '201':
            description: 'The rule has been deleted'
            schema:
              $ref: '#/definitions/Badge'
        '400':
          description: Invalid input
        '500':
          description: Internal server error
          
          
  /events:
    post:
      tags:
        - events
      summary: Add a new event
      description: ''
      operationId: addEvent
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: New event
          required: true
          schema:
            $ref: '#/definitions/Event'
      responses:
        '201':
            description: 'The event has been created'
            schema:
              $ref: '#/definitions/Event'
        '400':
          description: Invalid input
        '500':
          description: Internal server error
        
        
  /users/{userId}:
    get:
      tags:
        - users
      summary: Get user
      description: Returns user
      operationId: getUser
      parameters:
      - in: path
        name: userId
        type: integer
        required: true
        description: ID of the user to get
      produces:
        - application/json
      responses:
        '200':
          description: successful operation
          schema:
            type: object
            $ref: '#/definitions/User'
        '500':
          description: Internal server error
          
          
definitions:
  Rule:
    type: object
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
      appApiToken:
        type: string
  Event:
    type: object
    properties:
      userId:
        type: string
      apiToken:
        type: string
      name:
        type: string
  Badge:
    type: object
    properties:
      id:
        type: integer
        format: int64
  User:
    type: object
    properties:
      id:
        type: integer
        format: int64
      badges:
        type: array
        items: 
          type: object 
          $ref: '#/definitions/Badge'
  Application:
    type: object
    properties:
      apiToken:
        type: integer
        format: int64
      badges:
        type: array
        items: 
          type: object 
          $ref: '#/definitions/Badge'
      rules:
        type: array
        items: 
          type: object 
          $ref: '#/definitions/Rule'
      users:
        type: array
        items: 
          type: object 
          $ref: '#/definitions/User'    
  