# REST/src/main/java/io/swagger/api/RulesApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.4.1-SNAPSHOT).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package io.swagger.api;

import io.swagger.entities.RuleEntity;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;

import javax.validation.Valid;
import javax.validation.constraints.NotNull;
import java.util.List;
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2019-01-02T12:37:14.114Z")

@Api(value = "rules", description = "the rules API")
public interface RulesApi {

    @ApiOperation(value = "Add a new rule", nickname = "createRule", notes = "", response = RuleEntity.class, tags={ "rules", })
    @ApiResponses(value = { 
        @ApiResponse(code = 201, message = "The rule has been created", response = RuleEntity.class),
        @ApiResponse(code = 400, message = "Invalid input"),
        @ApiResponse(code = 500, message = "Internal server error") })
    @RequestMapping(value = "/rules",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.POST)
    ResponseEntity<RuleEntity> createRule(@ApiParam(value = "New rule" ,required=true )  @Valid @RequestBody RuleEntity body);


    @ApiOperation(value = "Delete rule", nickname = "deleteRule", notes = "Delete rule", response = RuleEntity.class, tags={ "rules", })
    @ApiResponses(value = { 
        @ApiResponse(code = 201, message = "The rule has been deleted", response = RuleEntity.class),
        @ApiResponse(code = 400, message = "Invalid input"),
        @ApiResponse(code = 500, message = "Internal server error") })
    @RequestMapping(value = "/rules",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.DELETE)
    ResponseEntity<RuleEntity> deleteRule(@ApiParam(value = "RuleEntity to be deleted" ,required=true )  @Valid @RequestBody RuleEntity body);


    @ApiOperation(value = "Get rules", nickname = "getRules", notes = "Returns list of rules", response = RuleEntity.class, responseContainer = "List", tags={ "rules", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "successful operation", response = RuleEntity.class, responseContainer = "List"),
        @ApiResponse(code = 500, message = "Internal server error") })
    @RequestMapping(value = "/rules",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<List<RuleEntity>> getRules(@NotNull @ApiParam(value = "", required = true) @Valid @RequestParam(value = "apitoken", required = true) String apitoken);


    @ApiOperation(value = "Update exisiting rule", nickname = "updateRule", notes = "", response = RuleEntity.class, tags={ "rules", })
    @ApiResponses(value = { 
        @ApiResponse(code = 201, message = "The rule has been updated", response = RuleEntity.class),
        @ApiResponse(code = 400, message = "Invalid input"),
        @ApiResponse(code = 500, message = "Internal server error") })
    @RequestMapping(value = "/rules",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.PUT)
    ResponseEntity<RuleEntity> updateRule(@ApiParam(value = "New rule" ,required=true )  @Valid @RequestBody RuleEntity body);

}